# contract/openapi.json
{
  "components": {
    "schemas": {
...
      "AttackGraph": {
        "additionalProperties": false,
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/ReplayFrame"
            },
            "title": "Frames",
            "type": "array"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "incident_id",
          "nodes",
          "edges",
          "frames"
        ],
        "title": "AttackGraph",
        "type": "object"
      },
...
      "Body_upload_case_api_cases_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_upload_case_api_cases_post",
        "type": "object"
      },
      "CaseCreated": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          }
        },
        "required": [
          "case_id"
        ],
        "title": "CaseCreated",
        "type": "object"
      },
      "CaseSource": {
        "enum": [
          "upload",
          "scenario"
        ],
        "title": "CaseSource",
        "type": "string"
      },
...
      "CaseSummary": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "contract_version": {
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "funnel": {
            "$ref": "#/components/schemas/Funnel"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parse_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParseReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "pipeline_version": {
            "title": "Pipeline Version",
            "type": "string"
          },
          "scenario_params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScenarioParams"
              },
              {
                "type": "null"
              }
            ]
          },
          "scenario_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Seed"
          },
          "source": {
            "$ref": "#/components/schemas/CaseSource"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          }
        },
        "required": [
          "case_id",
          "name",
          "source",
          "status",
          "pipeline_version",
          "funnel"
        ],
        "title": "CaseSummary",
        "type": "object"
      },
...
      "DismissalList": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "dismissals": {
            "items": {
              "$ref": "#/components/schemas/Dismissal"
            },
            "title": "Dismissals",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "dismissals"
        ],
        "title": "DismissalList",
        "type": "object"
      },
...
      "EntityProfile": {
        "additionalProperties": false,
        "properties": {
          "baseline": {
            "additionalProperties": true,
            "title": "Baseline",
            "type": "object"
          },
          "entity": {
            "$ref": "#/components/schemas/Entity"
          },
          "event_count": {
            "default": 0,
            "title": "Event Count",
            "type": "integer"
          },
          "observed": {
            "additionalProperties": true,
            "title": "Observed",
            "type": "object"
          }
        },
        "required": [
          "entity"
        ],
        "title": "EntityProfile",
        "type": "object"
      },
...
      "EventContext": {
        "additionalProperties": false,
        "properties": {
          "after": {
            "items": {
              "type": "string"
            },
            "title": "After",
            "type": "array"
          },
          "before": {
            "items": {
              "type": "string"
            },
            "title": "Before",
            "type": "array"
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "line_no": {
            "title": "Line No",
            "type": "integer"
          },
          "producing_signals": {
            "items": {
              "type": "string"
            },
            "title": "Producing Signals",
            "type": "array"
          }
        },
        "required": [
          "event",
          "before",
          "after",
          "line_no",
          "file_name"
        ],
        "title": "EventContext",
        "type": "object"
      },
      "EventsPage": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "title": "Events",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "case_id",
          "events",
          "total"
        ],
        "title": "EventsPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentDetail": {
        "additionalProperties": false,
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "title": "Claims",
            "type": "array"
          },
          "dismissals": {
            "items": {
              "$ref": "#/components/schemas/Dismissal"
            },
            "title": "Dismissals",
            "type": "array"
          },
          "incident": {
            "$ref": "#/components/schemas/Incident"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "title": "Links",
            "type": "array"
          },
          "score_breakdown": {
            "$ref": "#/components/schemas/ScoreBreakdown"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/Signal"
            },
            "title": "Signals",
            "type": "array"
          }
        },
        "required": [
          "incident",
          "claims",
          "score_breakdown"
        ],
        "title": "IncidentDetail",
        "type": "object"
      },
      "IncidentList": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/Incident"
            },
            "title": "Incidents",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "incidents"
        ],
        "title": "IncidentList",
        "type": "object"
      },
...
      "NaiveView": {
        "additionalProperties": false,
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/NaiveAlert"
            },
            "title": "Alerts",
            "type": "array"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "alerts"
        ],
        "title": "NaiveView",
        "type": "object"
      },
...
      "RevealResponse": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MatchResult"
            },
            "title": "Matches",
            "type": "array"
          },
          "scorecard": {
            "$ref": "#/components/schemas/Scorecard"
          },
          "truth": {
            "$ref": "#/components/schemas/GroundTruth"
          }
        },
        "required": [
          "case_id",
          "truth",
          "matches",
          "scorecard"
        ],
        "title": "RevealResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Forensic incident reconstruction API (contract v1.0.0)",
    "title": "221B",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/cases": {
      "post": {
        "operationId": "upload_case_api_cases_post",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_case_api_cases_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Case",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/dismissals": {
      "get": {
        "operationId": "dismissals_api_cases__case_id__dismissals_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DismissalList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dismissals",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/entities/{etype}/{eid}": {
      "get": {
        "operationId": "entity_api_cases__case_id__entities__etype___eid__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "etype",
            "required": true,
            "schema": {
              "title": "Etype",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "eid",
            "required": true,
            "schema": {
              "title": "Eid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Entity",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/events": {
      "get": {
        "operationId": "events_api_cases__case_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "host",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Host"
            }
          },
          {
            "in": "query",
            "name": "user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User"
            }
          },
          {
            "in": "query",
            "name": "ip",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ip"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/events/{event_id}": {
      "get": {
        "operationId": "event_context_api_cases__case_id__events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "context",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 25,
              "minimum": 0,
              "title": "Context",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventContext"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Context",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/incidents": {
      "get": {
        "operationId": "incidents_api_cases__case_id__incidents_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(incident|watchlist)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Incidents",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/incidents/{incident_id}": {
      "get": {
        "operationId": "incident_detail_api_cases__case_id__incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Incident Detail",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/incidents/{incident_id}/replay": {
      "get": {
        "operationId": "replay_api_cases__case_id__incidents__incident_id__replay_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttackGraph"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/naive": {
      "get": {
        "operationId": "naive_api_cases__case_id__naive_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NaiveView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Naive",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/progress": {
      "get": {
        "operationId": "progress_api_cases__case_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Progress",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/report.md": {
      "get": {
        "operationId": "report_api_cases__case_id__report_md_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/reveal": {
      "post": {
        "operationId": "reveal_api_cases__case_id__reveal_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevealResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reveal",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/summary": {
      "get": {
        "operationId": "summary_api_cases__case_id__summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Summary",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{case_id}/suspects": {
      "get": {
        "operationId": "suspects_api_cases__case_id__suspects_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspectList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suspects",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}